Contact page refactor w/ Tailwind
